#44 Rearranged itde stuff
