+++ pytest-backend/test/unit/test_itde_backend.py
@@ -1,20 +1,17 @@
 import pytest
 from exasol_integration_test_docker_environment.lib.test_environment.ports import Ports
 
-from exasol.pytest_backend.itde_config import (
-    ItdeOption,
-    ItdeOptionGroup,
-)
+import exasol.pytest_backend.config as config
 
 OPTIONS = (
-    ItdeOption(
+    config.Option(
         name="port",
         prefix="exasol",
         type=int,
         default=9999,
         help_text="Port to connect to",
     ),
-    ItdeOption(
+    config.Option(
         name="URL",
         prefix="exasol",
         type=str,
@@ -42,7 +39,7 @@ def test_get_pytest_option_name_for_option(option, expected):
 
 
 def test_help_of_option_without_default_value():
-    option = ItdeOption(
+    option = config.Option(
         name="port",
         prefix="exasol",
         type=int,
@@ -54,7 +51,7 @@ def test_help_of_option_without_default_value():
 
 
 def test_help_of_option_with_default_value():
-    option = ItdeOption(
+    option = config.Option(
         name="port",
         prefix="exasol",
         type=int,
@@ -116,8 +113,8 @@ BUCKETFS_OPTIONS = (
 @pytest.mark.parametrize(
     "group,expected",
     (
-        (ItdeOptionGroup(prefix="exasol", options=EXASOL_OPTIONS), "exasol"),
-        (ItdeOptionGroup(prefix="bucket", options=BUCKETFS_OPTIONS), "bucket"),
+        (config.OptionGroup(prefix="exasol", options=EXASOL_OPTIONS), "exasol"),
+        (config.OptionGroup(prefix="bucket", options=BUCKETFS_OPTIONS), "bucket"),
     ),
 )
 def test_option_group_prefix_property(group, expected):
@@ -129,30 +126,30 @@ def test_option_group_prefix_property(group, expected):
     "group,expected",
     (
         (
-                ItdeOptionGroup(prefix="exasol", options=EXASOL_OPTIONS),
-                (
-                    ItdeOption(
+            config.OptionGroup(prefix="exasol", options=EXASOL_OPTIONS),
+            (
+                config.Option(
                     prefix="exasol",
                     name="host",
                     type=str,
                     default="localhost",
                     help_text="Host to connect to",
                 ),
-                    ItdeOption(
+                config.Option(
                     prefix="exasol",
                     name="port",
                     type=int,
                     default=8563,
                     help_text="Port on which the exasol db is listening",
                 ),
-                    ItdeOption(
+                config.Option(
                     prefix="exasol",
                     name="username",
                     type=str,
                     default="SYS",
                     help_text="Username used to authenticate against the exasol db",
                 ),
-                    ItdeOption(
+                config.Option(
                     prefix="exasol",
                     name="password",
                     type=str,
@@ -178,7 +175,7 @@ class PyTestArgs:
     "group,env,cli,expected",
     (
         (
-                ItdeOptionGroup(
+            config.OptionGroup(
                 prefix="db",
                 options=(
                     {
@@ -189,12 +186,12 @@ class PyTestArgs:
                     },
                 ),
             ),
-                {},
-                PyTestArgs(),
-                {"port": 9999},
+            {},
+            PyTestArgs(),
+            {"port": 9999},
         ),
         (
-                ItdeOptionGroup(
+            config.OptionGroup(
                 prefix="db",
                 options=(
                     {
@@ -205,12 +202,12 @@ class PyTestArgs:
                     },
                 ),
             ),
-                {"DB_PORT": "7777"},
-                PyTestArgs(),
-                {"port": 7777},
+            {"DB_PORT": "7777"},
+            PyTestArgs(),
+            {"port": 7777},
         ),
         (
-                ItdeOptionGroup(
+            config.OptionGroup(
                 prefix="db",
                 options=(
                     {
@@ -221,9 +218,9 @@ class PyTestArgs:
                     },
                 ),
             ),
-                {"DB_PORT": "7777"},
-                PyTestArgs(db_port=8888),
-                {"port": 8888},
+            {"DB_PORT": "7777"},
+            PyTestArgs(db_port=8888),
+            {"port": 8888},
         ),
     ),
 )
